requirements.txt
discord
mysql-connector-python
